posts for specific user

+++ backend/src/controllers/posts.ts
@@ -2,12 +2,16 @@ import { RequestHandler } from "express";
 import PostModel from "../models/Post";
 import createHttpError from "http-errors";
 import mongoose from "mongoose";
+import { assertIsDefined } from "../util/assertisDefined";
 
 //getallposts
 export const getPosts: RequestHandler = async (req, res, next) => {
+  const authuser = req.session.userId;
+
   //RequestHandler Give varible type instead giving idividual types for parameters
   try {
-    const posts = await PostModel.find().exec(); //exec return promise
+    assertIsDefined(authuser)
+    const posts = await PostModel.find({userId : authuser}).exec(); //exec return promise
     res.status(200).json(posts); //200 equest was successful
   } catch (error) {
     next(error); //call error handler
@@ -17,8 +21,11 @@ export const getPosts: RequestHandler = async (req, res, next) => {
 //get specific post
 export const getPost: RequestHandler = async (req, res, next) => {
   const postId = req.params.postId;
+  const authuser = req.session.userId;
+
 
   try {
+    assertIsDefined(authuser)
     if (!mongoose.isValidObjectId(postId)) {
       throw createHttpError(400, "Invalid Post ID");
     }
@@ -26,6 +33,10 @@ export const getPost: RequestHandler = async (req, res, next) => {
     const post = await PostModel.findById(postId).exec();
     if (!post) {
       throw createHttpError(404, "Post not found");
+    }
+
+    if(!post.userId.equals(authuser)){
+      throw createHttpError(401, "Cannot acces this post");
     }
     res.status(200).json(post);
   } catch (error) {
@@ -46,9 +57,10 @@ export const createPosts: RequestHandler<
 > = async (req, res, next) => {
   const title = req.body.title;
   const text = req.body.text;
-  
+  const authuser = req.session.userId;
 
   try {
+    assertIsDefined(authuser)
     if (!title) {
       throw createHttpError(400, "Post must have a title"); //400 bad request
     }
@@ -57,6 +69,7 @@ export const createPosts: RequestHandler<
     }
 
     const newPost = await PostModel.create({
+      userId :authuser,
       title: title,
       text: text,
     });
@@ -82,8 +95,10 @@ export const updatePost: RequestHandler<
   const postId = req.params.postId;
   const title = req.body.title;
   const text = req.body.text;
-
+  const authuser = req.session.userId;
+  
   try {
+    assertIsDefined(authuser) 
     if (!mongoose.isValidObjectId(postId)) {
       throw createHttpError(400, "Invalid Post ID");
     }
@@ -97,6 +112,9 @@ export const updatePost: RequestHandler<
 
     if (!post) {
       throw createHttpError(404, "Post not found");
+    }
+    if(!post.userId.equals(authuser)){
+      throw createHttpError(401, "Cannot acces this post");
     }
 
     post.title = title;
@@ -112,8 +130,11 @@ export const updatePost: RequestHandler<
 
 export const deletePost: RequestHandler = async (req, res, next) => {
   const postId = req.params.postId;
+  const authuser = req.session.userId;
 
   try {
+    assertIsDefined(authuser) 
+
     if (!mongoose.isValidObjectId(postId)) {
       throw createHttpError(400, "Invalid Post ID");
     }
@@ -123,9 +144,12 @@ export const deletePost: RequestHandler = async (req, res, next) => {
       throw createHttpError(404, "Post not found");
     }
     // PostModel.findByIdAndRemove(postId);
+    if(!post.userId.equals(authuser)){
+      throw createHttpError(401, "Cannot acces this post");
+    }
     await PostModel.findByIdAndDelete(postId);
     res.sendStatus(204); //status dont send response it self json resoponsible for it in this case we use sendStatus for that
-  } catch (error) { 
+  } catch (error) {
     next(error);
   }
 };
